pyproject.toml
[tool.poetry]
name = "japan_postalcode_latlon_mapping"
version = "0.1.0"
description = ""
authors = ["RyoNak-billy <nakagamiryo0901@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
requests = "^2.32.2"
pandas = "^2.2.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
nbformat = "^5.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
